Its too slow :(. Break SourceMaps, remove some layers, enable loose mode for class to allow for mangle props
